backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Pillow==11.3.0
